[SWAGGER] Agendas + PropertiesFolders Controller OK, NOT TESTED

+++ controller/PropertiesFolders/PropertiesFoldersController.js
@@ -1,21 +1,21 @@
 const db = require('../../models/index.js');
 const propertiesFoldersTable = db['Properties_Folders'];
+const propertiesTable = db['Properties'];
+const clientFoldersTable = db['clientFolders'];
 
 const getPropertyFolder = async (req, res) => {
-
     try {
-        //  Récupération de l'utilisateur avec son id passé en paramètre d'URL
         const propertyFolder = await propertiesFoldersTable.findByPk(req.params.id);
-
-        res.status(200).send({
-            message: `Property-Folder ID : ${propertyFolder.id}`,
-            data: propertyFolder
-        })
-
+        if (propertyFolder !== null) {
+            res.status(200).send({
+                data: propertyFolder
+            })
+        } else {
+            res.status(404).send({
+                message: 'Property Folder not found'
+            })
+        }
     } catch (error) {
-
-        console.log(error);
-
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
             error: error.message
@@ -44,18 +44,17 @@ const getPropertyFolder = async (req, res) => {
 }
 const getPropertiesFolders = async (req, res) => {
     try {
-        //  Récupération de tous les utilisateurs
         const propertiesFolders = await propertiesFoldersTable.findAll();
-
-        //  Envoie de tous les utilisateurs
-        res.status(200).send({
-            message: 'Select all of properties-folders',
-            data: propertiesFolders
-        })
-
+        if (propertiesFolders && propertiesFolders.length !== 0) {
+            res.status(200).send({
+                data: propertiesFolders
+            })
+        } else {
+            res.status(404).send({
+                message: 'No propertiesFolders found'
+            })
+        }
     } catch (error) {
-        console.log(error);
-
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
             error: error.message
@@ -82,18 +81,36 @@ const getPropertiesFolders = async (req, res) => {
         // })
     }
 }
-const createPropertyFolder = async (req, res, next) => {
-
+const createPropertyFolder = async (req, res) => {
     try {
         const data = { ...req.body };
-        const newPropertyFolder = await propertiesFoldersTable.create(data);
-
-        next.
-            res.status(200).send({
-                message: 'Property-Folder created',
-                data: newPropertyFolder
+        const idProperty = data.idProperties;
+        const idPropertyFound = await propertiesTable.findByPk(idProperty);
+        if (idPropertyFound) {
+            const idClientFolders = data.idClientFolders;
+            const idClientFoldersFound = await clientFoldersTable.findByPk(idClientFolders);
+            if (idClientFoldersFound) {
+                const newPropertyFolder = await propertiesFoldersTable.create(data);
+                if (newPropertyFolder) {
+                    res.status(200).send({
+                        message: 'Property-Folder created',
+                        data: newPropertyFolder
+                    })
+                } else {
+                    res.status(400).send({
+                        message: 'PropertyFolder was not created'
+                    })
+                }
+            } else {
+                res.status(404).send({
+                    message: 'Client Folder not found'
+                })
+            }
+        } else {
+            res.status(404).send({
+                message: 'Property not found'
             })
-
+        }
     } catch (error) {
 
         console.log(error.message);
@@ -108,25 +125,46 @@ const modifyPropertyFolder = async (req, res) => {
 
     try {
 
-        const newData = { ...req.body };
         const idPropertyFolder = req.params.id;
-        const updatePropertyFolder = await propertiesFoldersTable.update(
-            newData,
-            {
-                where: {
-                    id: idPropertyFolder
+        const idPropertyFolderFound = await propertiesFoldersTable.findByPk(idPropertyFolder);
+        if (idPropertyFolderFound) {
+            const newData = { ...req.body };
+            const idProperty = newData.idProperties;
+            const idPropertyFound = await propertiesTable.findByPk(idProperty);
+            if (idPropertyFound) {
+                const idClientFolders = newData.idClientFolders;
+                const idClientFoldersFound = await clientFoldersTable.findByPk(idClientFolders);
+                if (idClientFoldersFound) {
+                    const updatePropertyFolder = await propertiesFoldersTable.update(
+                        newData,
+                        {
+                            where: { id: idPropertyFolder }
+                        })
+                    if (updatePropertyFolder[0] == 1) {
+                        res.status(200).send({
+                            message: 'Property-Folder updated'
+                        })
+                    } else {
+                        res.status(400).send({
+                            message: 'PropertyFolder not updated'
+                        })
+                    }
+                } else {
+                    res.status(404).send({
+                        message: 'Client folder not found'
+                    })
                 }
-            })
-        if (updatePropertyFolder[0] == 1) {
-            res.status(200).send({
-                message: 'Property-Folder updated'
+            } else {
+                res.status(404).send({
+                    message: 'Property not found'
+                })
+            }
+        } else {
+            res.status(404).send({
+                message: 'PropertyFolder not found'
             })
         }
-
     } catch (error) {
-
-        console.log(error);
-
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
             error: error.message
@@ -136,16 +174,26 @@ const modifyPropertyFolder = async (req, res) => {
 }
 const deletePropertyFolder = async (req, res) => {
     try {
-
-        const deletePropertyFolder = await propertiesFoldersTable.destroy({ where: { id: req.params.id } });
-        res.status(201).send({
-            message: 'Property-Folder deleted',
-            data: deletePropertyFolder
-        })
-
+        const idPropertyFodler = req.params.id;
+        const idPropertyFolderFound = await propertiesFoldersTable.findByPk(idPropertyFodler);
+        if (idPropertyFolderFound) {
+            const deletePropertyFolder = await propertiesFoldersTable.destroy({ where: { id: req.params.id } });
+            if (deletePropertyFolder !== 0) {
+                res.status(200).send({
+                    message: 'Property Folder deleted',
+                    data: deletePropertyFolder
+                })
+            } else {
+                res.status(400).send({
+                    message: 'Property Folder was not deleted'
+                })
+            }
+        } else {
+            res.status(404).send({
+                message: 'Property Folder not found'
+            })
+        }
     } catch (error) {
-        console.log(error);
-
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
             error: error.message
